requirements.txt
python-telegram-bot==21.3.0
python-dotenv==1.0.1
requests==2.32.3
